Tighten up member rule by requiring base to start at beginning of member.
Originally, I used "contains" to check whether base was the base
of member when propagating taint from members to base paths. That isn't
the full story though: we want to propagate taint only when the base is
a prefix of the member starting at index 0. This distinction seems to
not matter when edges only start at leaf fields, but better to be
precise.

This commit also introduces new macros for checking a particular access
path for the presence or absence of a tag which also introduces the
checked access path into the isAccessPath set. This allows us to propagate
privacy tags up to inner nodes that we care about checking without
generating all possible partial paths.